requirements.txt
attrs==25.3.0
beautifulsoup4==4.11.2
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
email-validator==2.3.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
google-auth==2.16.2
google-auth-oauthlib==1.0.0
greenlet==3.2.4
gunicorn==20.1.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.32
numpy==2.3.2
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
protobuf==3.20.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.28.2
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.35.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==2.2.3
wsproto==1.2.0
